Make a trade offer
It works, so the page shows all your items and a checkbox underneath to
choose the items to offer the person. I didnt put a submit button so it
wont submit anything yet. You access the offer trade by hitting the make
an offer link by the close project button. I couldnt get it to work with
a button so i used the link